BF-142 added 400 and 200 statuses for api method

+++ blackfriday/apps/landing/utils.py
@@ -3,10 +3,27 @@ from django.template.loader import render_to_string
 from django.conf import settings
 
 
-def render_landing():
+from apps.banners.models import Partner
+
+from .models import LandingLogo
+from .exceptions import NoContent
+
+
+def render_landing(raise_exception=True):
     path = os.path.join(settings.PROJECT_ROOT, 'landing')
     os.makedirs(path, exist_ok=True)
+    if not LandingLogo.objects.exists() and not Partner.objects.exists() and raise_exception:
+        raise NoContent
     with open(os.path.join(path, 'index.html'), 'w') as f:
         f.seek(0)
-        f.write(render_to_string('landing/landing.html', {'SITE_URL': settings.SITE_URL}))
+        f.write(
+            render_to_string(
+                'landing/landing.html',
+                {
+                    'SITE_URL': settings.SITE_URL,
+                    'partner_list': Partner.objects.all(),
+                    'logo_list': LandingLogo.objects.all(),
+                }
+            )
+        )
         f.truncate()
